Create Cart And Fix Product Service NEEDS FIX BUT AFTER AUTH AND FRONT

+++ Baqal.Application/DTOs/ProductDTO.cs
@@ -2,6 +2,7 @@
 using Baqal.Entities.Models;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -12,7 +13,8 @@ namespace Baqal.Application.DTOs
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+
+        public double Price { get; set; }
         public UnitType Unit { get; set; }
         public string Description { get; set; }
         public string ImageUrl { get; set; }
